NP-13634 Move components to allow splitting media forms
